LCCI.Announcements

Description

A module that holds examples for update models for different communication - types

Documentation

p :: Proposition

q :: Proposition

r :: Proposition

ingroup :: Atomic

outgroup :: Atomic

eventList :: [Event]

An infinite list with events, which is useful for zipping with or taking from. Numbering of this list starts at 1.

prtStr :: String -> Bool -> [Formula] -> String

worldIssue :: Ord a => [a] -> Issue a

Create an issue where the agent entertains which specific event occured

publicRaise :: String -> [Atomic] -> [Formula] -> (String, UpdateModel)

A function that create an update model for the public announcement of a question. Can be used for the public announcement of a declarative if only one formula is given. The formula will not check if the formulas given are actually declarative. The String is used as the name for the update model, and will be substituted by a default if left empty.

public :: [Atomic] -> Formula -> Int -> (String, UpdateModel)

A function that creates an update model for the public announcement of a given formula. The Int is for the maximum size of the model in which the formula will be announced, and can be ignored if the formula is declarative. It is used to calculate the resolutions of the formula. See publicRaise for more control over the asking of questions.

defaultPublicRaise :: [Atomic] -> [Formula] -> (String, UpdateModel)

A convience function for calling publicRaise without having to supply a string. Will always revert to the default.

privateRaise :: String -> [Atomic] -> [Atomic] -> [Atomic] -> [Formula] -> (String, UpdateModel)

Create an update model for the private announcement of a question. In this case, all the agents are aware that a question is being asked, but they are not all aware what this question is. This question is the inquisitive disjunction of the given list of formulas. The first list of Atomic agents are the ingroup, the group that knows which question is asked. The second are the outgroup, the group that does not know which question is asked, but wants to know, and the third are the agents that do not care about the question being asked. The rest of the parameters are as in publicRaise.

privateInform :: String -> [Atomic] -> [Atomic] -> [Atomic] -> [Formula] -> (String, UpdateModel)

Create an update model for the private announcement of some formula. In this case, all the agents are aware that some piece of information is given, but they are not all aware what this is. The first list of Atomic agents are the ingroup, the group that knows what is being told. The second are the outgroup, the group that does not know what is being told, but wants to know, and the third are the agents that do not care about the information. The list of formulas are all the possible pieces of information that can be shared. If this list has a length of 1, an extra event with precondition True will be added to the model. The rest of the parameters are as in publicRaise.

private :: [Atomic] -> [Atomic] -> [Atomic] -> Formula -> Int -> (String, UpdateModel)

A function that creates an update model for the private announcement of a given formula, where the rest of the group knows that some formula is annouced. See privateRaise for explanations on the groups, and public for an explanation of the Formula and Int

secretRaise :: String -> [Atomic] -> [Atomic] -> [Formula] -> (String, UpdateModel)

Create an update model for the secret announcement of a question. In this case, only a subset of the agents (the first parameter) are aware that a question is being asked, while the rest of the agents are not even aware that something is happening. the rest of the parameters are as in publicRaise.

secret :: [Atomic] -> [Atomic] -> Formula -> Int -> (String, UpdateModel)

A function that creates an update model for the secret announcement of a given formula, where the rest of the group is not aware that an announcement has been made. See secretRaise for a description of the first two parameters, and public for the description of the second two.