LCCI.Evaluation

Description

This module holds the code for the evaluation of LCCI formulas in models.

Documentation

compoundRelation :: World a => StaticModel a -> Program -> State a -> Set (State a)

Calculate all the states that are related to a given state by a program in the given model.

compoundRelation' :: World a => StaticModel a -> Program -> Relation a

Calculate the compound relation for a given model and program.

supports :: World a => StaticModel a -> State a -> Formula -> Bool

Check if the given formula is supported in the model and state.

productUpdate :: World a => StaticModel a -> UpdateModel -> StaticModel (a, Event)

Update the given StaticModel m with the given UpdateModel u.

updatedState :: World a => StaticModel a -> State a -> UpdateModel -> [Event] -> State (a, Event)

Calculate the updated state of s in m under application of events es from model u.