LCCI.Examples.Hexa

Documentation

w012 :: (Int, Int, Int)

The worlds in the Hex model

w021 :: (Int, Int, Int)

The worlds in the Hex model

w102 :: (Int, Int, Int)

The worlds in the Hex model

w120 :: (Int, Int, Int)

The worlds in the Hex model

w201 :: (Int, Int, Int)

The worlds in the Hex model

w210 :: (Int, Int, Int)

The worlds in the Hex model

ws :: Set (Int, Int, Int)

The set of worlds

a0 :: Proposition

The propositions in the Hex model

a1 :: Proposition

The propositions in the Hex model

a2 :: Proposition

The propositions in the Hex model

b0 :: Proposition

The propositions in the Hex model

b1 :: Proposition

The propositions in the Hex model

b2 :: Proposition

The propositions in the Hex model

c0 :: Proposition

The propositions in the Hex model

c1 :: Proposition

The propositions in the Hex model

c2 :: Proposition

The propositions in the Hex model

card :: Proposition -> String

Get the number of the card for the given proposition.

player :: Proposition -> Char

Get the identifier for the player in the given proposition.

v :: Valuation (Int, Int, Int)

The valuation of the propositional atoms

a :: Atomic

The "actions" in the Hex model (The knowledge relations)

b :: Atomic

The "actions" in the Hex model (The knowledge relations)

c :: Atomic

The "actions" in the Hex model (The knowledge relations)

ann_map :: StateMap (Int, Int, Int)

s :: Map Atomic (StateMap (Int, Int, Int))

The state maps for all the various agents.

hexa :: StaticModel (Int, Int, Int)

The actual model

ann_look :: (String, UpdateModel)

ann_events :: [Event]

hexa1 :: StaticModel ((Int, Int, Int), Event)

ws1 :: Set ((Int, Int, Int), Event)

hexa2 :: StaticModel (((Int, Int, Int), Event), Event)

ws2 :: Set (((Int, Int, Int), Event), Event)

hexa3 :: StaticModel ((((Int, Int, Int), Event), Event), Event)

ws3 :: Set ((((Int, Int, Int), Event), Event), Event)

e1 :: Event

e2 :: Event

e3 :: Event

ei :: StateMap Event

em :: Map Atomic (StateMap Event)

pre :: Map Event Formula

sub :: Map Event Substitution

swap_bill_carol :: UpdateModel

swap_events :: [Event]

swap_bill_carol' :: (String, UpdateModel)

Orphan instances

(Show a, Show b, Show c) => PrettyShow (a, b, c)

Instance details

Methods

prettyShow :: (a, b, c) -> String

(Ord a, Ord b, Ord c) => World (a, b, c)

Instance details