LCCI.Issue

Description

This module implements the structures from Inquisitive Semantics Information States, Issues, and State Maps.

Documentation

class Ord a => World a

A world in an LCCI model

type State a = Set a

An information state

state :: Ord a => [a] -> State a

A convience function for defining States. Creates a state with every world in the input list

type Issue a = Set (Set a)

A new issue. This will not garantee that an Issue is downward closed.

showIssue :: (PrettyShow a, Ord a) => Issue a -> String

A prettier way to show Issues

emptyIssue :: Issue a

Creates an empty issue, i.e. an issue that only contains the empty state

issue :: Ord a => [Set a] -> Issue a

Create an issue from a list of information states

powerset :: Ord a => Set a -> Set (Set a)

Create the powerset of a given set.

downwardClose :: Ord a => Set (Set a) -> Set (Set a)

Calculate the downward closure of a set of sets.

isDownwardClosed :: Ord a => Issue a -> Bool

Check to see if an issue is downward closed

alternatives :: Ord a => Issue a -> [Set a]

Find the alternatives in an issue

type StateMap a = Map a (Issue a)

A statemap for orderable objects

showStateMap :: (Ord a, PrettyShow a) => String -> String -> StateMap a -> String

Turn a statemap into a human readable string. The entries in the statemap are spereated by the string given in i.

showStateMaps :: (Ord a, PrettyShow a, PrettyShow b) => String -> Map b (StateMap a) -> String

Turn a map of statemaps into a human readable string. The elements in the statemap and between statemaps are seperated by the string i.