LCCI.Model

Description

This module implements both the static and update models of LCCI, including the Valuation

Documentation

type Valuation a = Proposition -> a -> Bool

The valuation for an IE-PDL model.

type ValuationMap a = Map a (Set Proposition)

A simple way to define a Valuation, by specifing which propositions are true in a world.

showValuationMap :: Show a => String -> ValuationMap a -> String

Turn the given ValuationMap into a string.

data StaticModel a

A StaticModel represents an IE-PDL model.

Constructors

StaticModel

Fields

worlds :: Set a

The worlds in this model.

valuation :: Valuation a

The valuation for this model. Should be a full function over worlds

staticStatemap :: Map Atomic (StateMap a)

The statemap belonging to the model. Should relate every world in worlds to some information state.

newtype Event

A type to represent events in the UpdateModel.

Constructors

Event Int

data UpdateModel

A type to represent an LCCI model, used to be combined with a StaticModel to evaluate Formulas with an update operator.

Constructors

UpdateModel

Fields

events :: Set Event

eventStatemap :: Map Atomic (StateMap Event)

precondition :: Map Event Formula

substitutions :: Map Event Substitution

Orphan instances

(World a, Ord b) => World (a, b)

Instance details