LCCI.Reduce

Description

This module allows for the reduction of formulas to equivalent formulas that - do not use the Update operator.

Documentation

reduce :: Int -> Formula -> Formula

Reduce a given LCCI formula to its IE-PDL equivalent

reduceStep :: Int -> Formula -> Formula

Set the next step of the reduction from LCCI to IE-PDL

tr :: UpdateModel -> Set Event -> Set Event -> Program -> Program

k :: UpdateModel -> Set Event -> Set Event -> Set Event -> Program -> Program