LCCI.Relation

Description

This module implements a simple model of Relations between information states.

Documentation

data Relation a

The relation type, used for relating information states to other relation states.

null :: Relation a -> Bool

Check if a given relation is empty.

empty :: Relation a

The empty relation, i.e., the relation that maps nothing to something else.

insert :: Ord a => Relation a -> State a -> State a -> Relation a

Add a relation between two information states.

fromList :: Ord a => [(State a, State a)] -> Relation a

Read a relation from a list of tuples.

fromStateMap :: World a => StateMap a -> Relation a

Read a relation from a given state map

keys :: Relation a -> [State a]

Get all the information states that this relation relates to others.

toList :: Relation a -> [(State a, State a)]

Turn this relation into a list of tuples

union :: Ord a => Relation a -> Relation a -> Relation a

Take the union of two relations.

compose :: Ord a => Relation a -> Relation a -> Relation a

Compose two relations together

reflexiveClosure :: Ord a => Relation a -> Relation a

Take the reflexive closure of this relation.

transitiveClosure :: Ord a => Relation a -> Relation a

Take the transitiveClosure of this relation.

mapDomain :: Ord b => (State a -> b) -> Relation a -> [b]

Map the given function f over the domain of the relation.

lookup :: Ord a => Relation a -> State a -> Set (State a)

Lookup to which states this relation relates a given information state. Returns an empty set if it maps it to none.

isValid :: Ord a => Relation a -> Bool

Check if the given relation adheres to all the properties that an IE-PDL relation should.

makeKeys :: Ord a => Relation a -> Relation a

Make all the keys for a given relation.

makeValid :: Ord a => Relation a -> Relation a

Make sure that the given relation adheres to all the properties it has to.

showRelation :: (PrettyShow a, Ord a) => Map Atomic (Relation a) -> String

Turn the given map of Relations into a string.