LCCI.Substitution

Description

A module that defines the substitutions for LCCI and functions for working with them. A substitution is a map from propositional atoms to formulas, that can be used to change the truth value of a proposition after the execution of an action.

Documentation

data Substitution

A type for the substitutions.

null :: Substitution -> Bool

Test to see if a substitution is empty

size :: Substitution -> Int

Calculate the size of a given substitution, i.e. how many propositions it maps to other formulas.

empty :: Substitution

The empty substitution.

insert :: Proposition -> Formula -> Substitution -> Substitution

Insert a new mapping into a substitution.

fromList :: [(Proposition, Formula)] -> Substitution

Create a substitution from a list of mappings.

toList :: Substitution -> [(Proposition, Formula)]

Transform the given substitution into a list of mappings.

domain :: Substitution -> [Proposition]

The domain of this substitution.

lookup :: Proposition -> Substitution -> Maybe Formula

Lookup a given proposition in the substitution. Returns Nothing if the proposition is not mapped to anything else.

apply :: Substitution -> Formula -> Formula

Apply the proposition to the substitution. The same as lookup, but returns the proposition in case it is not mapped to anything.

(!) :: Substitution -> Proposition -> Formula infixl 9

Lookup the given proposition in the substitution. Returns the proposition as a formula in case it is not mapped to anything.