LCCI.Syntax

Description

This module implements the Syntax of LCCI, and formulas for manipulating this syntax.

Documentation

data Atomic

An atomic program with no internal structure

atom :: String -> Atomic

Create a new atom from a string. The first letter will be lowercased

data Program

A full program

Constructors

Atom Atomic

Test Formula

Sequence [Program]

Choice [Program]

Iterate Program

data Proposition

A Logical Proposition

proposition :: String -> Proposition

Create a new Proposition from a string. The first letter will be uppercased

data Formula

A Formula in LCCI

Constructors

Prop Proposition

Bot

Top

Neg Formula

Quest Formula

And [Formula]

Or [Formula]

IOr [Formula]

Cond Formula Formula

BiCond Formula Formula

Modal Program Formula

IModal Program Formula

Update (String, UpdateModel) [Event] Formula

class Eq a => FlattenAble a

A datatype for structures that can be flattened, like programs and formulas

Minimal complete definition

flattenStep

flattenStep :: FlattenAble a => a -> a

Flatten a given structure by one step in the process.

flatten :: FlattenAble a => a -> a

Fully flatten the given structure.

class FlattenAble a => Simplify a

A datatype for structures that can be flattenend an simplified, like programs and formulas.

Minimal complete definition

simplifyStep

simplifyStep :: Simplify a => a -> a

Simplify the given structure by one step in the simplification process

simplify :: Simplify a => a -> a

Fully simplify the given structure.

expandStep :: Formula -> Formula

Expand the outer part of a formula by replacing the abbreviations. Also folds together conjunctions inside conjunctions etc.

expand :: Formula -> Formula

Expand a formula by replacing all of the abbreviations in them.

isDeclarative :: Formula -> Bool

Check if a given formula is declarative

resolutions :: Int -> Formula -> [Formula]

Give the resolutions of a given formula for models with at most Int worlds.