LCCI.Syntax.Pretty

Description

A module for making the reading and writing of formulas easier, by defining

operators and functions that are less efficient, but eaier to use.

Documentation

class PrettyShow a

A class for objects that can be shown in a "pretty" way, instead of just in a way that Haskell can read in.

Minimal complete definition

prettyShow

prettyShow :: PrettyShow a => a -> String

Show an object in a more human-readable manner

class FormulaLike a where

A class fo objects that can be cast into Formulas

Methods

asFormula :: a -> Formula

Cast an object into a formula

class ProgramLike a where

A class for objects that can be cast into programs.

Methods

asProgram :: a -> Program

Cast an object into a program

test :: FormulaLike a => a -> Program

Create a test program for a given formula

(.>) :: (ProgramLike a, ProgramLike b) => a -> b -> Program infixl 5

The sequence operator for programs

(.^) :: ProgramLike a => a -> Int -> Program infix 4

The power operator for programs

(.+) :: (ProgramLike a, ProgramLike b) => a -> b -> Program infixl 5

The choice operators for programs

iter :: ProgramLike a => a -> Program

Create the iterate program of a given program

polar :: FormulaLike a => a -> Formula

Create a polar question for a given formula

neg :: FormulaLike a => a -> Formula

Give the negation of a given formula

(/\) :: (FormulaLike a, FormulaLike b) => a -> b -> Formula infixl 5

The conjunction operator for formulas

(\/) :: (FormulaLike a, FormulaLike b) => a -> b -> Formula infixl 5

The disjunction operator for formulas

(\\/) :: (FormulaLike a, FormulaLike b) => a -> b -> Formula infixl 5

The inquisitive disjunction operator for formulas

(-->) :: (FormulaLike a, FormulaLike b) => a -> b -> Formula infix 6

The implication operator

(<->) :: (FormulaLike a, FormulaLike b) => a -> b -> Formula infix 6

Biimplication operator

knows :: (ProgramLike a, FormulaLike b) => a -> b -> Formula

The normal modality

entertains :: (ProgramLike a, FormulaLike b) => a -> b -> Formula

The inquisitive modality

wonder :: (ProgramLike a, FormulaLike b) => a -> b -> Formula

A function to quickly make the wonder modality

(@@) :: World a => StaticModel a -> UpdateModel -> StaticModel (a, Event) infix 1

The update procedure operator

(.@) :: World a => (StaticModel a, State a) -> (UpdateModel, [Event]) -> State (a, Event) infix 1

The updated state operator

(|=) :: World a => (StaticModel a, State a) -> Formula -> Bool infix 9

Supports operator