LCCI.Util

Description

This Module exports some utility functions that are in use in other parts of the project.

Documentation

permutate :: [[a]] -> [[a]]

Given a list of lists, calculate all the possible permutations of their elements. So it picks one element for each list, and then puts that in front of the permutations of the rest of the lists. This is the same as making a list comprehension that takes from each list, but for an unknown number of lists.

prettyShowWithParen :: PrettyShow a => String -> [a] -> String

Show a list of objects with parenthesis around it. The first argument is the string that is shown in between the elements of the list

class PrettyShow a

A class for objects that can be shown in a "pretty" way, instead of just in a way that Haskell can read in.

Minimal complete definition

prettyShow

prettyShow :: PrettyShow a => a -> String

Show an object in a more human-readable manner

prettyPrint :: PrettyShow a => a -> IO ()

Same as print, but using prettyShow instead of show.

takeUntil :: (a -> Bool) -> [a] -> [a]

Take from a list until a certain predicate p holds. This includes the first element for which the predicate holds.

match :: Eq a => [[a]] -> ([a], [[a]])

Calculate the matching starts for a list of lists, and return the mathing start together with the non-matching ends of the lists.