LCCI.Util.Tex

Description

A module that holds classes and functions for transforming a given datastructure into its (La)TeX representation. This module assumes things will be printed in mathmode, unless specifically mentioned otherwise or enabled by the resulting LaTeX code itself. It will also assume the excistens of the following LaTeX macros:

\lnot

Logical negation

\land

Logical conjunction

\lor

Logical disjunction

\lior

Inquisitive disjunction

\lif

Material conditional

\liff

Biimplication

\lac

Typesets the modal program operator. Takes one argument, which is the program that goes into the modal operator.

\liac

Typesets the inquisitive modal program operator. Takes one argument, which is the program that goes into the modal operator.

\lup

Typesets the update operator. Takes an optional argument (the name of the update model) and a required one (the (set of) events that is being executed.

\upd

Typesets in the font used for update model symbols. Takes one argument, the thing to be typeset in that font.

\pre

Typesets the function for the precondition. Takes one argument, the thing upon which the function is called.

\sub

Typesets the function for the substitution. Takes one argument, the thing upon which the function is called.

Documentation

class Texable a where

Convert a datastructure to a valid TeX representation of the structure. May use the commands given at the start of this module.

Minimal complete definition

Nothing

Methods

texWithParen :: String -> a -> String

Create a TeX representation of the given object, prepended with the given string on each line, put in braces.

texPrepend :: String -> a -> String

Create a TeX representation of the given object, prepend with the given string on each line

toTexWithParen :: a -> String

Create a TeX representation of the given subject, if needed within brackets

toTex :: a -> String

Create a TeX representation of a given object

enclose :: Texable a => String -> String -> String -> [a] -> String

Enclose the TeX representation of the list with the first and second argument, separated by the third. Useful for making lists and sets.

withParens :: Texable a => String -> [a] -> String

Get the TeX representation enclosed in parenthesis. The first argument is the separator for in between the elements.

asSet :: Texable a => [a] -> String